classifier_pipeline_medNLI/requirements.txt
pandas
tensorboard
torch
transformers
pytorch-lightning
pytorch-nlp
loguru